Fixed issue with address showing on musician index for annonymous users.
Modified the musician details page to look nicer if no albums or songs are found.

Added option to add musicians to a place at the time of a place's creation. This will move the musician from their former place to the newly created place.

Cleaned up some old code and comments.